pulling funcscope into its own file
